+++ src/Uri.php
@@ -225,7 +225,7 @@ class Uri implements UriInterface
         if (strlen($username) > 0) {
             return $this->with('userInfo', $this->constructString([
                 '%s%s'  => $username,
-                '%s:%s' => rawurlencode($password),
+                '%s:%s' => rawurlencode((string) $password),
             ]));
         }
 
@@ -409,7 +409,7 @@ class Uri implements UriInterface
         $formats = array_keys($components);
         $values = array_values($components);
         $keys = array_keys(array_filter($values, function ($value) {
-            return null !== $value && '' !== $value;
+            return $value !== null && $value !== '';
         }));
 
         return array_reduce($keys, function ($string, $key) use ($formats, $values) {