+++ src/UriParser.php
@@ -36,8 +36,15 @@ class UriParser
     /** Parsing mode that allows UTF-8 characters in some URI components */
     const MODE_UTF8 = 2;
 
-    /** Parsing mode that also converts international domain names to ascii */
+    /**
+     * Parsing mode that also converts international domain names to ascii
+     * @deprecated Use MODE_IDNA instead
+     * @see UriParser::MODE_IDNA
+     */
     const MODE_IDNA2003 = 4;
+
+    /** Parsing mode that also converts international domain names to ascii */
+    const MODE_IDNA = 4;
 
     /** @var array<string,string> List of methods used to assign the URI components */
     private static $setters = [
@@ -76,9 +83,9 @@ class UriParser
      *   query and fragment components of the URI. These characters will be
      *   converted to appropriate percent encoded sequences.
      *
-     * - `MODE_IDNA2003` also allows UTF-8 characters in the domain name and
+     * - `MODE_IDNA` also allows UTF-8 characters in the domain name and
      *   converts the international domain name to ascii according to the IDNA
-     *   2003 standard.
+     *   standard.
      *
      * @param int $mode One of the parsing mode constants
      */
@@ -183,7 +190,7 @@ class UriParser
     {
         if (preg_match('/^[\\x00-\\x7F]*$/', $hostname)) {
             return $hostname;
-        } elseif ($this->mode !== self::MODE_IDNA2003) {
+        } elseif ($this->mode !== self::MODE_IDNA) {
             throw new \InvalidArgumentException("Invalid hostname '$hostname'");
         }