+++ CHANGES.md
@@ -2,10 +2,11 @@
 
 ## Unreleased
 
+  * Add support for PHP 8.2
+  * Fix passing nulls to string functions
+  * DEPRECATED `UriParser::MODE_IDNA2003` in favor of `UriParser::MODE_IDNA`
   * Minor bundled autoloader improvements
-  * Minor improvements to the build process
   * Address minor code quality issues in tests
-  * Add tests builds to support PHP 8.2
   * Update CI to use Github Actions
 
 ## v2.1.0 (2017-06-30) ##